Add support for overriding default config values
This is most useful with configs with `!env:json` that expect a list
rather than an object.  This is kind of a hack, but it's what we've got.

For example, `web_server.registered_clients` is not expecting a map but
a list. If you copied the previous value, the web-server service
wouldn't start and give this error:
`cfg.login.registeredClients.map is not a function`