Update to docker-exec-websocket-* 3.0.0
Update docker-exec-websocket-client and docker-exec-websocket-server.
These should work under Node v14, and are required for Node v16. They
fix an issue when the `docker exec` command is reading and writing data
to the container.

The test was using Buffer.compare(), which returns 0 when the buffers
are equal, 1 when the first argument sorts first, and -1 when the second
argument sorts first. This means the test was confirming the two buffers
were _not_ equal, since 2015 or so.