Add support for checks in the github badge API
When taskcluster uses the `reporting: checks-v1` it doesn't create a
global github status anymore but a check per task. The taskcluster
github badge API used to only check for said status which would thus not
be there and return "No data yet".

This commit adds support for the github service to request the checks
associated with the reference we're requesting a badge for. I chose to
combine those results with a status in case it's also present. I don't
think this can happen in real situations but it doesn't cost anything.

Since checks are per task instead of task group, here are the rules for
the badge state:

- If any check is a failure, the state is marked as a failure.
- If no checks have a failure but there's at least one that pending,
  state is marked as pending
- Otherwise all checks are green, mark the state as success.
- If no checks are present, this keeps the same behavior as before.

A quick note on implementation: We are using the app ID to find which
check is coming from the taskcluster bot because that's what github
provides in its checks API. That differs from the status filtering which
is using the user id of the github app.

Fixes #5139