Fix broken tests on win2012r2 CI
Python 2 replaced with python 3 on worker which required updating
one test that ran a python 2 script. Also mozilla-build is no longer
installed on the CI worker, so needed to migrate test that used msys
bash to use cygwin bash instead. Not sure if it still holds much
value, was originally created for https://bugzil.la/1360539.