fix(github): use checks API to fulfill latest endpoint
The /repository/<owner>/<repo>/<branch>/latest endpoint relies on the
Github `statuses` API to retrieve the link to the task group of a
commit. But many repos are opting into the `checks-v2` reporting (which
will likely become the default eventually). Practically speaking, this
means the badge links for these projects are busted.

This PR tweaks the API to use the Github `checks` API to build a URL to
the `check` page if a project has opted into checks reporting.

This is backwards compatible as this endpoint will continue to look at
the `status` API first, and only fallback to `checks` if status was
empty.

Fixes #5364