Fix the `latest` endpoint in the Github service when using checks
I had implemented this endpoint under the assumption that the
`findTCChecks` function was calling Github's `check-suites` API, when in
fact it was using the `check-runs` API.

Since check-runs points to individual tasks, I grab the run with the
smallest id, which should correlate to the Decision task. Even if it
doesn't, it's not a big deal as all tasks are easily accessible on the
side.

Fixes #5395