docker-worker: set a timeout when downloading artifact
If the connection to download the docker image hangs for whatever
reason, we could wait indefinitely, and because the maxRunTime timer
isn't set up by that point, we'd only stop when hitting the task's
deadline and failing to reclaim it.
Instead, time out the connection if it stays idle for 60 seconds, so we
can properly retry the download.