feat(worker-manager): extra check before worker is terminated on terminateAfter time exceeded.
This is an attempt to reduce the amount of workers that are being
stopped even after they were registered and started claiming work.
In some conditions, especially for Azure workers, scan loops can take
10-20 minutes. And since all records are being fetched in memory once,
it may happen that worker record was updated. We simply reload the
record before finally removing worker.