feat(github): Cancellation of builds filters for the event type
If new event has 'push' type, only builds for this SHA and 'push' types
would be cancelled.
If it is a pull_request, then ['opened', 'synchronize'] would be treated
as same event type, and the rest would be individual.
So if there are three events: 'pull_request.opened',
'pull_request.synchronize', 'pull_request.assigned', then not all of
them will be auto-cancelled, but only the ones that much same event type