Rework the interactive feature for generic worker
Instead of sending commands line by line and reading responses the same
way, present the pty to the client through a websocket. This allows for
interactive applications to run as well as displaying things as they
would be in a normal terminal emulator. This includes support for
resizing which means the PTY will adjust to the window size of the user.

This changes how interactive commands work on a fundamental level.
Instead of running one command that everyone attaches to, this spawns
one command per connection. That means that a command is spawned when a
websocket opens instead of when the task starts running.

To be able to handle resizing, we need to distinguish resize requests
coming from the client. Since we only have one websocket, we use the
first byte of every message to distinguish whether the remaining of the
message is data to be put on the PTY or a size request.