feat(g-w): Allow livelog exposed port to be configured
By default it will remain `0` and so port would be allocated
dynamically.
But for local development environment where generic-worker runs inside
the container, and `livelog` binary is executed within same container,
assigning ports dynamically would make it hard to expose it using static
docker-compose.yml file. This way we provide predefined port that will
be used to expose livelog, which then could be accessed on the host and
in the UI.

--wip-- [skip ci]