Also change the version in generic worker CI files on release
The CI file for generic-worker was moved in 13c21c268bfc62726064ece7740b7ad92d70e21b
but the release script wasn't updated to accomodate that. One of the
command needs the version to match the current release and was
automatically updated by the release script. After the path change, that
version fell out of sync leading to CI failures because some linker
argument was wrong and it wasn't setting the `revision` parameter.

Fixes #6513