Queue internals refactoring Part 2/3 (#6554)
* feat(db): initial migration to new data structure
* feat(db): using base64 original column instead of separate payload jsonb
* make claim and deadline queue work
* expire pending tasks
* adding claimed, resolved, and deadline tasks into tables
* removing taskcluster-lib-azqueue
* deadline resolver is working
* claimed queue works
* resolved tasks functions
* fixes for test:meta
* adding indexes and fixing queries
* poll pending is a function
* remove unnecessary queueMock helper
* When task is resolved we drop message from claim queue.
This also introduces notifications for `task_pending` and
`task_resolved` channels.
Those would be later used to replace constant polling with fetch+listen.
* Improved methods descriptions and added tests
* Add documentation about internal workings of queue and task lifecycle
* Speed up DependencyResolver, we can iterate faster
* Updated changelog message
* Added comment
* Fixes after review.
Unique taskId+runId ensured on claimed task table to avoid having
multiple records for the same run.
* Remove count from config and use constant instead. This number shouldn't
be configurable per deployment
* Removing "Azure" references in code and updating queue API description
* Remove deadline message as soon as task is resolved.
Dependencies will be updated by DependencyTracker once resolved message
lands in resolved queue, so we don't have to update dependencies at
deadline for this task anymore.
* Revert back to pending queues might have at least one record.
From PR comment:
Historically, queues have been "over-estimations", so for example a task
being in the pending queue meant that the task might actually be
pending. But a task that was not in the pending queue was definitely not
pending. This is a pretty common distributed-algorithm design, but was
necessary with azure queues since it's impossible to delete a message
from a queue - or know if a message exists in a queue - until that
message is visible and comes to the top of the queue.
* Adding downgrade migration and patched azure_* methods
Existing methods would be forced to use new table structure.
To assist downgrade migration new columns were added. They will be
dropped in upcoming migration.
* Added tests for up/down migrations
Data should be preserved when migrating both ways, and running processes
should be safe to write to the currently available tables
* Include expiring now messages just in case
* Introduce unique taskId<>runId for pending tasks
When task is created multiple times it will result in multiple entries
in pending queue.
With unique key on those fields we have the ability to enforce that
there will be no duplicate entries for the same task and run.
* This test relied on the fact that api response was cached
This is no longer the case and we can ensure that extra messages do not
appear in pending tasks by using unique keys on the table.
* Apply suggestions from code review

Co-authored-by: Matt Boris <92693437+matt-boris@users.noreply.github.com>