Allow specifying `0` as a value for `onExitStatus.purgeCaches`
I was trying to abuse the caching system to download a binary that I was
developing on another machine to run in a task. Took me by surprise that
it didn't allow automatically purging the cache when the command
succeeded. As far as I can tell, docker-worker did not have such a
limit and I don't see why there should be one anyway.