Don't close the websocket when invalid UTF-8 is output by an interactive session
There's 2 parts to this. First, the worker was using a `TextMessage` for
websockets which would assume the message was valid UTF-8. This ended up
being the main issue, it would just close the websocket when it saw
invalid UTF-8.

The second part was the DECODER (which wasn't used in version 2 of
interactive sessions until now), which was spamming the console with
errors when it received invalid UTF-8.

Fixing both of those allows one to see invalid UTF-8 output in their
interactive sessions.

I kept backwards compatibility with generic-worker which are still
sending `TextMessage` messages by checking that the received data is a
string and kept the old behavior in that case.

Fixes #6848