Add a proper `TERM` environment variable to interactive sessions
The `TERM` environment variable is used to indicate the capabilities of
the current terminal emulator. Applications use it to know whether they
can use some features or not. An example of that is tmux which looks for
the `clear` capability which is missing from the default `TERM` assigned
by bash (`dumb`).
Thankfully this was fairly easy as `hterm` is a supported terminal.