Fix docker worker shell UI to set row/cols properly
It looks like this is broken on first load, due to the second `resize` call, as well as on resize (onTerminalResize and DockerExecClient.resize's arg order doesn't match).

There's an initial `this.client.resize.apply` call which is correct...and seems like it doesn't need to exist...but this is working fine so it seems best not to try to remove it.