feat: introduce new queue APIs for batch task requests
This is done according to RFC 189¹.
In few words, the motivation is to drastically reduce the number of
queue API calls done by decision tasks, since they sometimes need
the status for hundreds of tasks.

During implementation, I found out that the Go client code generation
struggled with `patternProperties`, which are needed to directly
return an object whose keys would be the task ids.

After speaking with @lotas, we decided that using an array
would be fine for now to avoid going down this rabbit hole.

[1] taskcluster/taskcluster-rfcs#189