Ensure 'taskGroup' is always defined when fetching scopes for submitting an action task.
This is in support of making actions like "Cancel All" work on task groups that have been created by an Action Task (see https://github.com/mozilla/firefox-translations-training/issues/250).

In an ideal world I'd like to understanding how `task.decisionTask` fits into things a bit more, and see if it could be removed...but that's more work than I'm willing to put in at the moment. I think this is a fairly safe fix to take in the meantime.