feat(queue): resolved tasks are removed from pending queue
This issue existed for a long time and queue refactoring made it
apparent that tasks that were considered pending, were in fact not.
When task is resolved (cancelled), queue service only removed it form
the claimed list, but record in pending queue was left as is.
This was not an issue for queue service, because it was also checking
real task status at claim time, but worker-manager only saw total
numbers. And seeing non-zero pending count resulted in creation of new
instances that would start and then do nothing because no tasks were
actually pending.