User `userdel` instead of `deluser`
The idea is the same as for the `adduser` removal, `deluser` is debian
specific and this change allows running the worker on non debian
derivative distributions.

As with `useradd`, the flags between the two match pretty well. In fact,
the only change is `--remove-home` that gets translated to `-r`.

```
-r Files in the user's home directory will be removed along with the home directory itself and the user's mail spool. Files located
in other file systems will have to be searched for and deleted manually.
```

That behavior is the exact same as with `deluser --remove-home`.