fix(queue)!: Emit pulse message to 'task-exception' regardless of retry state
BREAKING CHANGE: Pulse events now sent for task runs in the exception
state that are automatically retried.

Previously, we were only emitting a message if the run was not being
retried. This is a bug as we need to be able to track all runs of a
task.

I moved the publish logic above the call to `task-pending` as this
matches the order of events (first the old run had an exception, then
the new run is pending).