feat(generic-worker): set a TASK_WORKDIR env
It's often the case that we need to find an absolute path to various
things in a task's payload (e.g in the env or command). This is
currently difficult with generic-worker because the task user's home
directory is not known when creating the task definition.

We've done some hacky workarounds in the run-task script like this:
https://github.com/taskcluster/taskgraph/blob/d4e75591a98c12e0fbb0060190860f23b7c39d0b/src/taskgraph/run-task/run-task#L1260

But it occurred to me that if generic-worker could just set the env,
none of this would be necessary.