Force all task users to share the same subuid map
This fixes the cache issues we've been having with podman with #7149
applied.

The core idea here is that on podman, caches permissions for UIDs
(outside of 1000) are mapped to the user subuid range on the host. So UID 0 becomes
100000, 1 becomes 100001 and so on. When remounting the cache in the
next container, the reverse operation happens to get the right UID in
the container, 100000 on the host becomes 0 in the container.

However, by default, users do not share subuid ranges, `adduser` will
always use a new subuid range (100000 for the first user, +64k for the
second and so on). Which means that the mapping wouldn't work on the
second job and the cache would just get mounted as `nobody:nogroup`.

This commit changes the user creation so that every task user shares the
same subuid range.

Fixes #7404