Return a struct containing information about what d2g did
This will allow users of the conversion functions to interact with the
container created by the converted command. It's necessary to allow the
interactive feature to find that container back.

Note that the container name is necessary since it's randomly generated.

This changes the behavior of D2G a bit as the container used to only be
named if there were artifacts but is now named at all time.