Add the ability to have a wait command to interactive
This is preparation work for the interactive feature to drop the user in
a container as we need to wait for said container to be started.

The main point here is that because docker and podman don't allow
waiting for a container to be in a state if the container doesn't exist,
we have to allow the wait command to fail. Hence the retry loop.