Drop the user in the container in interactive sessions with D2G
This wires everything from the commits before to change the interactive
command when d2g is involved. In that case, instead of running bash,
it'll exec bash into the container and use that pty instead.

It uses `docker wait` to wait for the container to be running.
Note that that command fails if the container isn't created yet (i.e if
the image isn't done downloading). There's unfortunately nothing we can
do about this without overcomplicating the wait command. So instead it
uses the fact that the wait command is allowed to fail for 20s by the
interactive feature. We could change that time to be higher if it shows
to not be enough.