Change the waiting model to have a "ready command" instead
Turns out that podman is better than docker at having useful commands
and `docker wait` does not support `--condition`. Because of that we
can't just have one command wait for a container to be ready.

So instead, I replaced it with a command that returns whether or not the
container is ready and it runs 20 times, once a second until it is
ready, driven by a loop in go.

The string interpolation in the commands now looks scarier than before
but both `ContainerName` comes directly from d2g without any unsanitized
user input.