WriteToFileAsJSON now first writes to a temp file and then moves the file in-place.
The goal here is that if you are replacing a file in a known good state with
another version of that file, and you hit an issue when writing the new version
to disk, the old version will not be replaced, and you should not end up with a
corrupted file. Note, this still isn't foolproof, since a filesystem move could
potentially cross file system boundaries, and thus imply copy/delete semantics.