fix(g-w): better detection of the docker bridge ip address
In the presence of more than one address on the docker bridge network,
we were assuming the gateway could be found on the first network, which
is not necessarily the case.  Instead, we now get all gateways from
`docker network inspect`, and select the first ipv4 one (ipv6 addresses
would need more massaging to work in URLs, so we skip them).