Azure certificates fixes (#7674)
* feat(w-m): Remove bundled root certificates

* feat(w-m): Automate azure certificates loading

Adding a script that downloads certificates from the list that is later
used for verification purposes in Azure provider.
This also updates the README.md to list what those certificates are and
when they expire.

* Remove expired Azure certs

* feat(w-m): Include current Azure certificates

Readme is updated to explain current state.
certificates.json tells which urls to download by `download-certs.js`
script.
Information about expiration date is added to the README.md

* Removed expired root certificate