Add github API throttling
This was added before but got removed by
af70daceb82a3ae3ccf6176f76636b69f8c19aff because the plugin has a
default that adds a 1 second delay between each request writing to the
github API.

This commit basically re-adds the plugin but also sets that minimum time
to 50ms instead of the default of 1s. I've been running this in
production for a few weeks now and I haven't seen anything weird. Task
groups with 500+ tasks that used to get me rate limited hard are now
not even hitting the primary rate limit but the checks are still pretty
much immediately available on github so I think it's good enough as it
is.

Fixes #7646