Properly reap the proxy process when stopping the feature
On posix systems, when killing a child process, the parent process has
to call `wait(2)` on it or the child process will be left in a "zombie"
state until the parent exits, at which point PID 1 will do that.

For generic worker, it means that every time it starts a proxy, it leaks
one zombie as it's never waiting for it until it gets restarted which in
some cases could be never (headless multiuser, insecure engine).

By simply calling `Wait` on the proxy process after killing it, it now
ends up calling `wait(2)` on it, reaping it. We simply ignore the actual
return status and print an error in case something went horribly wrong.

Fixes #5786