Use prometheus as plugin in Monitor library
Initially it seemed like a good idea to have separate metrics library
but upon further investigation and implementation it became clear that
it would be easier to delegate handling of metrics to the monitor
library. Prometheus is exposed as a plugin.
MonitorManager.registerMetric() is introduced to register metrics in the
same way as log types are being registered with .register()

This should simplify maintenance and configuration