feat(ui): Fix signed url issue without credentials
This happened because user session is being initialized twice.
First time it is initialized right after successful callback from
external provider, and at this point credentials are not being issued.
Second time it is initialized when AuthController detects that
credentials are missing, so it fetches them and updates user session
with valid credentials.

So between those two events error happened and tried to sign a url
without valid credentials.