workers: remove support for the dind and dockerSave docker-worker features
dind was not actually implemented in d2g as far as I can tell.  Tasks
using it in firefox-ci migrated away over the years as things broke due
to the docker daemon on the docker-worker images being too old.

dockerSave seems to be unused, and its implementation in generic-worker
appears not to work, so rather than try to fix a feature we don't need,
stop pretending it's there.