feat(w-m): update launch config always
Initial implementation was not updating configuration if the "main" part
wasn't changed. But that left "workerManager" object that was supposed
to be dynamic be excluded from updates, and the only way to update it
was to make sure new launchConfigId is being created.
Since we wanted to make "workerManager" part dynamic  (can be updated
without changing launchConfigId), we update now configuration even when
the main part isn't change, but workerManager part could change.