feat(api): collecting api calls metrics
Renamed `path`->`name` because urls can get quite big, and come with
greater variety. This explodes tracked buckets for each. Keeping it down
to the api method name avoids having too many entries, plus it is
probably safer not to store req.path as is to avoid unexpected
injections of malicious stuff