Replace minio with localstack for local development
Minio decided that their business model was now that people shouldn't be
using it, so let's agree with that and use something else.

The other alternative I knew about was garage but it's way more complex
to deploy in a single node than localstack which is made exactly for
that kind of use case.