Port over the minimal code for the Task Profiler implementation (#7376)
* Port over the minimal code for the Task Profiler implementation

This is the initial code needed for converting Taskcluster tasks and
live logs into the Firefox Profiler format. It is taken from:

https://github.com/gregtatum/taskcluster-tools

The goal with this PR is not to have a full working piece of code, but
to port over the minimal amount of code, and get TypeScript to have no
issues with the code. I added TODOs across the code while porting for
things that will need to be handled for the real implementation.

* Fix or disable lint errors

* Add a changelog

* Add TypeScript to the package.json