Fix early return in resolveTask sometimes skipping all-resolved dependents
This fixes a regression from commit cd4a7ae7f82 where the code was
changed to a for loop (from a .query()), which means that `return`
changed semantic from skipping one dependent to skipping all remaining
dependents.

This bug is very easily reproducible with the following graph:

- task A - fails
- task B - all-completed depends on A
- task C - all-resolved depends on A
- task B's ID is lexicographically ordered before C's
  (`get_dependent_tasks` returns task ordered by ID)

In that case, neither B or C would get scheduled even though C should
be.

Fixes #7829