Increase the send buffer size on websocktunnel clients
Without this, the client has a 4kB default buffer size which it sends
and then awaits an ACK before sending another buffer... (that's still
the case with this but at least it's sending a maximum of 4MB of data
every time).

I tested this with 100ms delay both sides on websocktunnel and it went
from ~500 lines/s to 60k lines/s with `- head -c 100000000 /dev/urandom
| base64 && sleep 30` while catching up which should be more than
enough.