Use proper URL encoding in the go client for path parameters
The go client was using `QueryEscape` to escape path parameters,
however, that is not the correct function to escape paths (`PathEscape`
as the name implies is better suited for that).

`QueryEscape` follows the `application/x-www-form-urlencoded` spec,
which encodes more than what paths require but also replaces spaces by
`+`. The over-encoding isn't an issue since the server would decode that,
but the space replacement means that for example, a secret with a space
in it could never get retrieved by the go client (as it would request
`a+b` instead of `a b`.)

This behavior was very visible when uploading artifacts with a space in
their name as they'd end up stored with a `+` instead. It gets worse
when you had an artifact named `a b` and `a+b` as it would upload the
second one on top of the first. All services in the client were affected
but since the taskcluster CLI uses its own URL builder for API calls
(where it calls `strings.ReplaceAll(url.QueryEscape(v), "+", "%20")` to
work around this issue), the only product really affected by this change
is generic-worker.

I added a simple test using the secrets service because this is the
easiest way I figured out how to test the behavior without a complicated
setup.

This is a breaking change because anyone relying on space in artifacts
being replaced by a `+` will have to remove their workarounds for it.

It's also worth noting that this brings the go client in line with the
other clients (js, py, rs) which were already all doing the right thing.

Fixes #7765